python/agent-hub/keyframe-agent/pyproject.toml
[project]
name = "keyframe-agent"
version = "0.0.1"
description = "Keyframe Agent for generating Tintin-style keyframes"
authors = [{name = "Your Name", email = "your.email@example.com"}]
requires-python = ">=3.12"
dependencies = [
    "mofa-ai @ file:///root/mofa-euterpe/python",
    "requests>=2.31.0",
    "python-dotenv==1.0.1",
    "pydantic==2.10.6",
]

[project.scripts]
keyframe-agent = "keyframe_agent.agent.main:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "keyframe-agent"
version = "0.0.1"
description = "Keyframe Agent for generating Tintin-style keyframes"
authors = ["Your Name <your.email@example.com>"]
packages = [{include = "keyframe_agent"}]

[tool.poetry.dependencies]
python = "^3.12"
mofa-ai = { path = "/root/mofa-euterpe/python", develop = true }
requests = ">=2.31.0"
python-dotenv = "1.0.1"
pydantic = "2.10.6"

[tool.poetry.scripts]
keyframe-agent = "keyframe_agent.agent.main:main"